Model the SearchTermScore
The `SearchTermScore` is the final output of one request to this app -
it represents the value of the search term paired with it's popularity
score.

One of the requirements of the application is the ability to retrieve
the results of the previous searches from the database, so this entity
has a mapping to database.