Model the SearchTerm
The search system works as follows:
user provided search term is modified by appending a sentiment keyword
to it. Positive search is represented by a `rocks` keyword, and a
negative search by a `suck` keyword.

The constructed phrase is then used to perform the search by a search
provider (could be GH Issues, Twitter, G+, FB...).