Model the Search Service
This is the main Search service of the application, which makes use of
the concrete search providers.

Basic interaction in easy - just call the `Search::search()` method with
a string representation of a search term. The response is the
`SearchTermScore` object. The Search service handles the positive and
negative sentiment searches automatically.